feat: compatibility with k8s
OK so running JupyterHub on Kubernetes is super hard. The
zero-to-jupyterhub ("z2jh") project runs the KubeSpawner, which is otherwise
undocumented. We deployed z2jh to observe how it works. In particular,
we added a "jupyterhub" service account to this plugin. This allows the
KubeSpawner to monitor the k8s cluster for pods creation/deletion.